+++ src/assets/soundDb.ts
@@ -1,26 +1,28 @@
+// Start from 01
+
 const p = 'modules/pf2e-graphics/assets/library'
 const database = {
 	lightning_bolt: {
 		impact: {
-			0: `${p}/sounds/ovani-sounds/Magic/Shock/Lightning Bolt Impact A.wav`,
-			1: `${p}/sounds/ovani-sounds/Magic/Shock/Lightning Bolt Impact B.wav`,
-			2: `${p}/sounds/ovani-sounds/Magic/Shock/Lightning Bolt Impact C.wav`,
+			'01': `${p}/sounds/ovani-sounds/Magic/Shock/Lightning Bolt Impact A.wav`,
+			'02': `${p}/sounds/ovani-sounds/Magic/Shock/Lightning Bolt Impact B.wav`,
+			'03': `${p}/sounds/ovani-sounds/Magic/Shock/Lightning Bolt Impact C.wav`,
 		},
 	},
 	bow: {
-		0: `${p}/sounds/gamedev-market/BowWhistleShot1.wav`,
-		1: `${p}/sounds/gamedev-market/BowWhistleShot2.wav`,
-		2: `${p}/sounds/gamedev-market/BowWhistleShot3.wav`,
-		3: `${p}/sounds/gamedev-market/BowWhistleShot4.wav`,
-		4: `${p}/sounds/gamedev-market/BowWhistleShot5.wav`,
-		5: `${p}/sounds/gamedev-market/BowWhistleShot6.wav`,
-		6: `${p}/sounds/gamedev-market/BowWhistleShot7.wav`,
+		'01': `${p}/sounds/gamedev-market/BowWhistleShot1.wav`,
+		'02': `${p}/sounds/gamedev-market/BowWhistleShot2.wav`,
+		'03': `${p}/sounds/gamedev-market/BowWhistleShot3.wav`,
+		'04': `${p}/sounds/gamedev-market/BowWhistleShot4.wav`,
+		'05': `${p}/sounds/gamedev-market/BowWhistleShot5.wav`,
+		'06': `${p}/sounds/gamedev-market/BowWhistleShot6.wav`,
+		'07': `${p}/sounds/gamedev-market/BowWhistleShot7.wav`,
 	},
 	crossbow: {
-		0: `${p}/sounds/gamedev-market/Crossbow Shot A.wav`,
-		1: `${p}/sounds/gamedev-market/Crossbow Shot B.wav`,
-		2: `${p}/sounds/gamedev-market/Crossbow Shot C.wav`,
-		3: `${p}/sounds/gamedev-market/Crossbow Shot D.wav`,
+		'01': `${p}/sounds/gamedev-market/Crossbow Shot A.wav`,
+		'02': `${p}/sounds/gamedev-market/Crossbow Shot B.wav`,
+		'03': `${p}/sounds/gamedev-market/Crossbow Shot C.wav`,
+		'04': `${p}/sounds/gamedev-market/Crossbow Shot D.wav`,
 	},
 }